fix(ci): python coverage shard — oversized shards and incomplete hosted handoff fake broke focused tests
